+++ brutils/__init__.py
@@ -46,6 +46,9 @@ from brutils.cpf import (
 from brutils.email import is_valid as is_valid_email
 
 # IBGE Imports
+from brutils.ibge.municipality import (
+    get_code_by_municipality_name,
+)
 from brutils.ibge.uf import (
     convert_code_to_uf,
 )
@@ -172,4 +175,5 @@ __all__ = [
     "is_valid_voter_id",
     # IBGE
     "convert_code_to_uf",
+    "get_code_by_municipality_name",
 ]