+++ brutils/__init__.py
@@ -44,6 +44,9 @@ from brutils.cpf import (
 
 # Email Import
 from brutils.email import is_valid as is_valid_email
+from brutils.ibge.municipality import (
+    get_municipality_by_code,
+)
 
 # IBGE Imports
 from brutils.ibge.uf import (
@@ -172,4 +175,5 @@ __all__ = [
     "is_valid_voter_id",
     # IBGE
     "convert_code_to_uf",
+    "get_municipality_by_code",
 ]